refactored tooltipData from parent, ClinicalChart to child InfoBox
